backend/requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
classify-imports==4.2.0
click==8.0.3
colorama==0.4.4
coverage==7.2.2
cryptography==38.0.4
cssselect2==0.7.0
Django==4.0.5
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-decorator-include==3.0
django-rest-auth==0.9.5
django-staticfiles==1.2.1
djangorestframework==3.14.0
execnet==1.9.0
Flask==2.0.2
Flask-Cors==3.0.10
future==0.18.2
html5lib==1.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.9.2
MarkupSafe==2.0.1
oscrypto==1.3.0
packaging==21.2
pep8==1.7.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
py==1.11.0
py-linq==1.3.0
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
pyparsing==2.4.7
PyPDF3==1.0.6
pyTelegramBotAPI==4.7.0
pytest==6.2.5
pytest-cache==1.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-pep8==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.6.0
reportlab==3.6.12
requests==2.27.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
vk==2.0.2
vk-api==11.9.7
webencodings==0.5.1
Werkzeug==2.0.2
xhtml2pdf==0.2.8
